# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 知行创境教育创作平台 API
  version: 1.0.0
  description: |
    基于《知行创境》教育创作平台 V2.0 PRD 整理的后端接口 OpenAPI 文档。
    基础路径为 `/api/v1`，认证方式为 Bearer Token。
servers:
  - url: /api/v1
    description: Default API Server
...
paths:
...
  /digital-humans/{digitalHumanId}/equipments:
    post:
      tags: [DigitalHuman]
      summary: 绑定博物馆装备
      operationId: addDigitalHumanEquipment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DigitalHumanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindEquipmentRequest'
      responses:
        '200':
          description: 绑定成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DigitalHumanEquipment'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /digital-humans/{digitalHumanId}/equipments/{equipmentId}:
    delete:
      tags: [DigitalHuman]
      summary: 删除已绑定装备
      operationId: deleteDigitalHumanEquipment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DigitalHumanId'
        - $ref: '#/components/parameters/EquipmentId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /community/exhibitions:
    get:
      tags: [Community]
      summary: 获取社区展厅列表
      operationId: getCommunityExhibitions
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: grade
          schema:
            type: string
        - in: query
          name: theme
          schema:
            type: string
        - in: query
          name: tag
          schema:
            type: string
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [latest, hot, likes]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 社区展厅分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedExhibitionSummary'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /community/exhibitions/{exhibitionId}:
    get:
      tags: [Community]
      summary: 获取社区展厅详情
      operationId: getCommunityExhibitionDetail
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      responses:
        '200':
          description: 社区展厅详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExhibitionViewerData'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /community/exhibitions/{exhibitionId}/like:
    post:
      tags: [Community]
      summary: 点赞展厅
      operationId: likeExhibition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Community]
      summary: 取消点赞
      operationId: unlikeExhibition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /community/exhibitions/{exhibitionId}/favorite:
    post:
      tags: [Community]
      summary: 收藏展厅
      operationId: favoriteExhibition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Community]
      summary: 取消收藏
      operationId: unfavoriteExhibition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /community/exhibitions/{exhibitionId}/share:
    post:
      tags: [Community]
      summary: 记录分享
      operationId: shareExhibition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareExhibitionRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /community/exhibitions/{exhibitionId}/comments:
    get:
      tags: [Community]
      summary: 获取评论列表
      operationId: getExhibitionComments
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 评论列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CommentSection'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Community]
      summary: 发表评论
      operationId: createExhibitionComment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '200':
          description: 评论成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Comment'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /community/exhibitions/{exhibitionId}/feature:
    post:
      tags: [Community]
      summary: 教师推荐到社区首页
      operationId: featureExhibition
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExhibitionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureExhibitionRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ExhibitionId:
      in: path
      name: exhibitionId
      required: true
      schema:
        type: integer
        format: int64
...
    DigitalHumanId:
      in: path
      name: digitalHumanId
      required: true
      schema:
        type: integer
        format: int64
    EquipmentId:
      in: path
      name: equipmentId
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    SuccessResponse:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 0
            message: ok
            data: {}
            requestId: 4f9b6e24e2ef4f4f9d6c
    ErrorResponse:
      description: 业务异常
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: 40100
                message: 未登录或 token 失效
                data: null
                requestId: 4f9b6e24e2ef4f4f9d6c
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
        requestId:
          type: string
      required: [code, message, requestId]
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 40001
        message:
          type: string
          example: 参数错误
        data:
          nullable: true
        requestId:
          type: string
      required: [code, message, requestId]
    UserRole:
      type: string
      enum: [student, teacher, parent, admin, visitor]
...
    ExhibitionStatus:
      type: string
      enum: [draft, published, archived]
    ExhibitionVisibility:
      type: string
      enum: [private, class, public]
...
    SimpleUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        role:
          $ref: '#/components/schemas/UserRole'
        name:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
          format: uri
...
    ExhibitionStats:
      type: object
      properties:
        viewCount:
          type: integer
        likeCount:
          type: integer
        favoriteCount:
          type: integer
        commentCount:
          type: integer
    ExhibitionSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
        title:
          type: string
        coverUrl:
          type: string
          format: uri
        summary:
          type: string
        status:
          $ref: '#/components/schemas/ExhibitionStatus'
        visibility:
          $ref: '#/components/schemas/ExhibitionVisibility'
        groupName:
          type: string
        ownerId:
          type: integer
          format: int64
        author:
          $ref: '#/components/schemas/SimpleUser'
        latestVersionNo:
          type: integer
        publishedVersionNo:
          type: integer
        stats:
          $ref: '#/components/schemas/ExhibitionStats'
        tags:
          type: array
          items:
            type: string
    ExhibitionDetail:
      allOf:
        - $ref: '#/components/schemas/ExhibitionSummary'
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            collaborators:
              type: array
              items:
                $ref: '#/components/schemas/ExhibitionMember'
    ExhibitionMember:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        role:
          type: string
          enum: [owner, editor, viewer]
        name:
          type: string
        avatarUrl:
          type: string
          format: uri
        joinedAt:
          type: string
          format: date-time
...
    CanvasConfig:
      type: object
      properties:
        width:
          type: integer
        height:
          type: integer
        background:
          type: string
        zoom:
          type: number
          format: double
    ExhibitionElement:
      type: object
      properties:
        componentType:
          type: string
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
        props:
          type: object
          additionalProperties: true
    ExhibitionRenderData:
      type: object
      properties:
        canvasConfig:
          $ref: '#/components/schemas/CanvasConfig'
        elements:
          type: array
          items:
            $ref: '#/components/schemas/ExhibitionElement'
...
    StoryTimelineItem:
      type: object
      properties:
        anchorCode:
          type: string
        startSecond:
          type: integer
        endSecond:
          type: integer
        content:
          type: string
    DigitalHumanEquipment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        slotCode:
          type: string
        anchorCode:
          type: string
        displayOrder:
          type: integer
        resourceId:
          type: integer
          format: int64
        resourceTitle:
          type: string
        museumName:
          type: string
        resourceSnapshot:
          type: object
          additionalProperties: true
    DigitalHuman:
      type: object
      properties:
        id:
          type: integer
          format: int64
        exhibitionId:
          type: integer
          format: int64
        name:
          type: string
        avatar2dUrl:
          type: string
          format: uri
        model3dUrl:
          type: string
          format: uri
        persona:
          type: string
        voiceType:
          type: string
        storyScript:
          type: string
        storyTimeline:
          type: array
          items:
            $ref: '#/components/schemas/StoryTimelineItem'
        equippedItems:
          type: array
          items:
            $ref: '#/components/schemas/DigitalHumanEquipment'
...
    BindEquipmentRequest:
      type: object
      required: [slotCode, museumResourceId]
      properties:
        slotCode:
          type: string
        museumResourceId:
          type: integer
          format: int64
        displayOrder:
          type: integer
        anchorCode:
          type: string
...
    SubmissionReview:
      type: object
      properties:
        id:
          type: integer
          format: int64
        reviewer:
          $ref: '#/components/schemas/SimpleUser'
        score:
          type: number
          format: double
        commentText:
          type: string
        commentAudioUrl:
          type: string
          format: uri
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        exhibitionId:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/SimpleUser'
        parentCommentId:
          type: integer
          format: int64
          nullable: true
        content:
          type: string
        mentionUsers:
          type: array
          items:
            $ref: '#/components/schemas/SimpleUser'
        createdAt:
          type: string
          format: date-time
    CommentSection:
      type: object
      properties:
        teacherReviews:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionReview'
        studentComments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    CreateCommentRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
        parentCommentId:
          type: integer
          format: int64
          nullable: true
        mentionUserIds:
          type: array
          items:
            type: integer
            format: int64
    ShareExhibitionRequest:
      type: object
      required: [channel]
      properties:
        channel:
          type: string
          example: wechat
    FeatureExhibitionRequest:
      type: object
      required: [featured]
      properties:
        featured:
          type: boolean
        featuredReason:
          type: string
...
    ExhibitionViewerData:
      type: object
      properties:
        exhibition:
          $ref: '#/components/schemas/ExhibitionDetail'
        renderData:
          $ref: '#/components/schemas/ExhibitionRenderData'
        digitalHuman:
          $ref: '#/components/schemas/DigitalHuman'
        teacherReviews:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionReview'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
...
    PagedExhibitionSummary:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/ExhibitionSummary'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...